@@ -56,7 +56,7 @@ func main() {
 
 func run() error {
 	var absMediaFile, mediaType string
-	var mediaFile interface{}
+	var mediaFile any
 	var isSeek bool
 
 	exitCTX, cancel := signal.NotifyContext(context.Background(), syscall.SIGINT, syscall.SIGTERM)
