@@ -124,10 +124,12 @@ func subsAction(screen *FyneScreen) {
 }
 
 func playAction(screen *FyneScreen) {
-	var mediaFile, subsFile interface{}
+	var mediaFile, subsFile any
 	w := screen.Current
 
-	screen.PlayPause.Disable()
+	fyne.Do(func() {
+		screen.PlayPause.Disable()
+	})
 
 	if screen.cancelEnablePlay != nil {
 		screen.cancelEnablePlay()
