@@ -17,7 +17,7 @@ var (
 
 // ServeTranscodedStream passes an input file or io.Reader to ffmpeg and writes the output directly
 // to our io.Writer.
-func ServeTranscodedStream(w io.Writer, input interface{}, ff *exec.Cmd, ffmpegPath, subs string, seekSeconds int) error {
+func ServeTranscodedStream(w io.Writer, input any, ff *exec.Cmd, ffmpegPath, subs string, seekSeconds int) error {
 	// Pipe streaming is not great as explained here
 	// https://video.stackexchange.com/questions/34087/ffmpeg-fails-on-pipe-to-pipe-video-decoding.
 	// That's why if we have the option to pass the file directly to ffmpeg, we should.
@@ -43,6 +43,8 @@ func ServeTranscodedStream(w io.Writer, input interface{}, ff *exec.Cmd, ffmpegP
 			vf = fmt.Sprintf("subtitles='%s',format=yuv420p", subs)
 		}
 	}
+
+	vf = "scale='min(1920,iw)':min'(1080,ih)':force_original_aspect_ratio=decrease," + vf
 
 	cmd := exec.Command(
 		ffmpegPath,
