@@ -54,7 +54,9 @@ MIT
 		}()
 	})
 	checkversion := widget.NewButton(lang.L("Check version"), func() {
-		go checkVersion(s)
+		go fyne.Do(func() {
+			checkVersion(s)
+		})
 	})
 
 	s.CheckVersion = checkversion
